+++ authority-portal-frontend/src/app/pages/authority-connector-list-page/authority-connector-list-page/authority-connector-list-page.component.ts
@@ -6,7 +6,12 @@ import {Store} from '@ngxs/store';
 import {ConnectorOverviewEntryDto} from '@sovity.de/authority-portal-client';
 import {GlobalStateUtils} from 'src/app/core/global-state/global-state-utils';
 import {sliderOverNavigation} from 'src/app/core/utils/helper';
-import {getConnectorsTypeClasses} from 'src/app/core/utils/ui-utils';
+import {
+  getConnectorStatusInnerCircleClasses,
+  getConnectorStatusOuterRingClasses,
+  getConnectorStatusText,
+  getConnectorsTypeClasses,
+} from 'src/app/core/utils/ui-utils';
 import {FilterBarConfig} from 'src/app/shared/components/common/filter-bar/filter-bar.model';
 import {HeaderBarConfig} from 'src/app/shared/components/common/header-bar/header-bar.model';
 import {
@@ -40,6 +45,9 @@ export class AuthorityConnectorListPageComponent implements OnInit, OnDestroy {
   filterBarConfig!: FilterBarConfig;
 
   getConnectorsTypeClasses = getConnectorsTypeClasses;
+  getConnectorStatusOuterRingClasses = getConnectorStatusOuterRingClasses;
+  getConnectorStatusInnerCircleClasses = getConnectorStatusInnerCircleClasses;
+  getConnectorStatusText = getConnectorStatusText;
 
   constructor(
     private store: Store,